+++ centered-cinnamon-dock@mostlynick3/CHANGELOG.md
@@ -26,3 +26,4 @@
  * Harden script to prevent race condition crashes on extension load/Cinnamon load
  * Correct size and trigger zone of side panels, and alignment with side indicators
  * Add and set new default auto hide mode, hide if focused window covers full screen
+ * Add welcome modal to new users, offering easy method to enable initial panels